// admin-clients/libs/shared/venues/data-access/standard-venue-tpls/src/lib/models/vm-item-type.enum.ts
export enum VenueTemplateItemType {
    sector = 'sect',
    row = 'row',
    seat = 'seat',
    notNumberedZone = 'nnz',
    aisle = 'aisle',
    notNumberedZoneStatusCounter = 'nnzSC',
    notNumberedZoneBlockingReasonCounter = 'nnzBRC',
    notNumberedZoneSessionPackCounter = 'nnzSPC',
    notNumberedZoneQuotaCounters = 'nnzQC'
}